Add .contains to support use as a set
This is to support zero-sized types (ZSTs) as values.  Note that at
the deepest leaves, we don't allocate space for such values, so `get`
will return None though the key is in the trie.

There's a reasonable chance we'll want to revisit how ZSTs are handled
in this code, but for now this at least gives us sensible set
behavior.

We also expand the tests here to cover some cases that previously
behaved badly without this change.